requirements.txt
python-dotenv
groq
sentence-transformers
joblib
pandas
scikit-learn
pdfplumber
openpyxl
pikepdf
streamlit